pyproject.toml
[project]
name = "pcb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "albumentations>=2.0.8",
    "font-manager>=0.1.0",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "kaggle>=1.8.3",
    "kagglehub>=0.4.0",
    "numpy>=2.4.1",
    "onnx>=1.20.1",
    "onnxruntime>=1.23.2",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "streamlit>=1.52.2",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "ultralytics>=8.3.252",
]
